requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beanie==1.27.0
click==8.1.7
dnspython==2.7.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
lazy-model==0.2.0
MarkupSafe==3.0.1
motor==3.6.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.31.1
